+++ src/Humanizer/NumberToWordsExtension.cs
@@ -20,6 +20,8 @@ namespace Humanizer
             {
                 case "ar":
                     return ToArabicWords(number);
+                case "fa":
+                    return ToFarsiWords(number);
                 default:
                     return ToEnglishWords(number);
             }
@@ -187,6 +189,57 @@ namespace Humanizer
             }
 
             return result.Trim();
+        }
+
+        private static string ToFarsiWords(int number)
+        {
+            if (number < 0)
+                return string.Format("منفی {0}", ToFarsiWords(-number));
+
+            if (number == 0)
+                return "صفر";
+
+            var farsiHundredsMap = new[]
+            {
+                "صفر", "صد", "دویست", "سیصد", "چهارصد", "پانصد", "ششصد", "هفتصد", "هشتصد", "نهصد"
+            };
+            var farsiTensMap = new[] 
+            {
+                 "صفر", "ده", "بیست", "سی", "چهل", "پنجاه", "شصت", "هفتاد", "هشتاد", "نود"
+            };
+            var farsiUnitsMap = new[]
+            {
+                 "صفر", "یک", "دو", "سه", "چهار", "پنج", "شش", "هفت", "هشت", "نه", "ده", "یازده", "دوازده", "سیزده", "چهارده", "پانزده", "شانزده", "هفده", "هجده", "نوزده"
+            };
+            var farsiGroupsMap = new Dictionary<int, Func<int, string>> 
+            { 
+                {(int)Math.Pow(10, 9), n => string.Format("{0} میلیارد", ToFarsiWords(n)) },
+                {(int)Math.Pow(10, 6), n => string.Format("{0} میلیون", ToFarsiWords(n)) },
+                {(int)Math.Pow(10, 3), n => string.Format("{0} هزار", ToFarsiWords(n)) },
+                {(int)Math.Pow(10, 2), n => farsiHundredsMap[n]}
+            };
+            var parts = new List<string>();
+            foreach (var group in farsiGroupsMap.Keys)
+            {
+                if (number / group > 0)
+                {
+                    parts.Add(farsiGroupsMap[group](number / group));
+                    number %= group;
+                }
+            }
+
+            if (number >= 20)
+            {
+                parts.Add(farsiTensMap[number / 10]);
+                number %= 10;
+            }
+
+            if (number > 0)
+            {
+                parts.Add(farsiUnitsMap[number]);
+            }
+
+            return string.Join(" و ", parts);
         }
     }
 }