+++ release_notes.md
@@ -1,3 +1,9 @@
+###v1.16.1 - 2014-04-03
+  - [#121] (https://github.com/MehdiK/Humanizer/pull/121) : Added Farsi support to DateTime and NumberToWords
+	- Added Farsi support to ToQuantityExtensions
+  
+[Commits](https://github.com/MehdiK/Humanizer/compare/v1.15.1...v1.16.1)
+  
 ###In Development
   - [#120](https://github.com/MehdiK/Humanizer/pull/120): Added translation for culture 'de'